Simplify route handlers
