+++ node-app/templates/clusterrole.yaml
@@ -3,7 +3,7 @@ apiVersion: rbac.authorization.k8s.io/v1
 metadata:
   name: {{ include "node-app.fullname" . }}-role
   labels:
-        {{- include "node-app.labels" . | nindent 4 }}
+    {{- include "node-app.labels" . | nindent 4 }}
 rules:
   - apiGroups:
       - ""
