+++ node-app/templates/podsecuritypolicy.yaml
@@ -1,0 +1,18 @@
+kind: PodSecurityPolicy
+apiVersion: policy/v1beta1
+metadata:
+  name: {{ include "node-app.fullname" . }}-psp
+  labels:
+    {{- include "node-app.labels" . | nindent 4 }}
+spec:
+  privileged: false
+  runAsUser:
+    rule: 'RunAsAny'
+  seLinux:
+    rule: 'RunAsAny'
+  supplementalGroups:
+    rule: 'RunAsAny'
+  fsGroup:
+    rule: 'RunAsAny'
+  volumes:
+    - 'hostPath'
