+++ package.json
@@ -24,6 +24,7 @@
   "dependencies": {
     "@hapi/glue": "^7.0.0",
     "@hapi/hapi": "18.4.0",
-    "@kubernetes/client-node": "^0.11.0"
+    "@kubernetes/client-node": "^0.11.0",
+    "hapi-pino": "^6.3.0"
   }
 }
