+++ routes/root/kubernetes-client-factory.js
@@ -1,0 +1,12 @@
+const kubernetes = require('@kubernetes/client-node');
+
+function buildClient() {
+    const config = new kubernetes.KubeConfig();
+    config.loadFromCluster();
+
+    return config.makeApiClient(kubernetes.CoreV1Api);
+}
+
+module.exports = {
+    create: buildClient
+};
